% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structured-package.R
\docType{package}
\name{structured-package}
\alias{structured}
\alias{structured-package}
\title{structured: A Tool To Create Structured Projects}
\description{
A tool to create structured projects. The projects can be created either
interactively using the RStudio Project Templates feature, or
programmatically using the \link{structure_project} function exported by this
package.
}
\author{
\strong{Maintainer}: Giuseppe Tomio \email{g.tinti.tomio@gmail.com}

}
\keyword{internal}
