% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure-project.R
\name{structure_project}
\alias{structure_project}
\title{Create a structured project}
\usage{
structure_project(
  project_path,
  project_level = c("Basic", "Intermediate"),
  analysis_format = c("Script", "R Markdown (Simplified)", "R Markdown", "Shiny")
)
}
\arguments{
\item{project_path}{Path to the new project.}

\item{project_level}{Project level. One of "Basic" or "Intermediate".}

\item{analysis_format}{Analysis format. One of "Script", "R Markdown (Simplified)", "R Markdown" or
"Shiny".}
}
\description{
Create a structured project from a given template.
}
