% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure-project.R
\name{structure_project}
\alias{structure_project}
\title{Create a structured project}
\usage{
structure_project(
  path,
  level = c("Basic", "Intermediate"),
  type = c("Script", "R Markdown (Simplified)", "R Markdown", "Targets", "Shiny")
)
}
\arguments{
\item{path}{Path to the new project.}

\item{level}{Project level. One of "Basic" or "Intermediate".}

\item{type}{Project type. One of "Script", "R Markdown (Simplified)", "R Markdown",
"Targets" or "Shiny".}
}
\description{
Create a structured project from a given template.
}
