% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure-project.R
\name{structure_project}
\alias{structure_project}
\title{Create a structured project}
\usage{
structure_project(
  path,
  level = c("Basic", "Intermediate"),
  type = c("Script", "R Markdown (Simplified)", "R Markdown", "Targets", "Shiny",
    "None"),
  open = FALSE
)
}
\arguments{
\item{path}{Path to the new project.}

\item{level}{Project level. One of "Basic" or "Intermediate". Default to "Basic".}

\item{type}{Project type. One of "Script", "R Markdown (Simplified)", "R Markdown",
"Targets", "Shiny" or "None". Default to "Script".}

\item{open}{Open option. If \code{TRUE} and on RStudio open the project in a new session, if
\code{TRUE} and not on RStudio just set the working directory to the new project
directory. If \code{FALSE} (default) do nothing.}
}
\value{
Return the project path invisibly.
}
\description{
Create a structured project from a given template.
}
\examples{
\dontrun{
structure_project("my-cool-project", level = "Basic", type = "Script")
}

}
